lab: rename storybook configs to mjs
