+++ packages/eds-lab-react/.storybook/manager.mjs
@@ -1,6 +1,6 @@
 import { addons } from '@storybook/manager-api'
 
-import { theme } from './theme'
+import { theme } from './theme.mjs'
 
 addons.setConfig({
   theme,
